requirements.txt
# 核心依赖
requests>=2.31.0
PyYAML>=6.0
click>=8.1.0
rich>=13.0.0
streamlit>=1.40.1
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.24.0
